Update BannerViewController  with banner refresh implementation
In MoPub IOS SDK, server-side triggered banner refresh does not update the MPAdView. The only way around this at the moment is to control your own refresh by calling 'loadAd' yourself on a timer. Updated sample project with a quick implementation of this
Created a constant kSelfBannerRefreshTimer for you to use in Constants.h